IOS-00025 Refactor under a shared interface IOS-00025C Project file organization
