+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scores/Child/MovementDetails/MovementDetailsView+Chart.swift
