+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scores/ScoresView.swift
@@ -1,160 +1,9 @@
 //
-//  ScoresFeature.swift
+//  ScoresView.swift
 //  StepTrackerTCA
 //
-//  Created by Sebastian Sciuba on 14/03/2025.
+//  Created by Sebastian Sciuba on 15/03/2025.
 //
-
-protocol ScoresFeatureServices {
-    func bestSession(from sessions: [any WorkoutSessionProtocol]) -> (any WorkoutSessionProtocol)?
-}
-
-
-final class DefaultScoresFeatureServices: ScoresFeatureServices {
-    
-    func bestSession(from sessions: [any WorkoutSessionProtocol]) -> (any WorkoutSessionProtocol)? {
-        return sessions.max { Double($0.value) ?? 0 < Double($1.value) ?? 0 }
-    }
-    
-}
-
-import ComposableArchitecture
-import Foundation
-
-@Reducer
-struct ScoresFeature {
-    
-    // MARK: - Dependencies
-    
-    let services: ScoresFeatureServices
-    
-    // MARK: - Livecycle
-    
-    init(service: ScoresFeatureServices = DefaultScoresFeatureServices()) {
-        self.services = service
-    }
-    
-    // MARK: - Reducer
-    
-    var body: some Reducer<State, Action> {
-        Reduce { state,action in
-            switch action {
-                // MARK: - Action
-                
-            case .updateWorkout:
-                let selectedWorkout = state.selectedWorkout
-                state.bestResult = services.bestSession(from: selectedWorkout.movements.flatMap { $0.sessions })
-                return .none
-                
-                // MARK: - View Action
-            case .view(.viewDidAppear):
-                return .run { send in
-                    await send(.updateWorkout)
-                }
-            case let .view(.openExerciseInfo(movement)):
-                state.destination = .openInfo(MovementInfoFeature.State(movement: movement))
-                return .none
-            case let .view(.openMovementDetails(movement, sessions)):
-                state.destination = .showDetails(MovementDetailsFeature.State(movement: movement, sessions: sessions))
-                return .none
-                
-                
-                // MARK: - Destination
-            case .destination:
-                return .none
-            }
-        }
-        
-        .ifLet(\.$destination, action: \.destination)
-    }
-}
-
-import ComposableArchitecture
-import Foundation
-
-/// Defines the actions available in `ScoresFeature`, handling data fetching,
-/// state updates, and user interactions.
-extension ScoresFeature {
-    @CasePathable
-    enum Action: ViewAction {
-        // MARK: - Actions
-        
-        case updateWorkout
-        
-        // MARK: - View Actions
-        
-        /// Handles view-related actions.
-        case view(View)
-        
-        /// Defines user interactions within the view.
-        enum View {
-            
-            /// Triggered when the view appears.
-            case viewDidAppear
-            
-            /// Opens the exercise details for a given `MovementType`.
-            /// This action is triggered when the user selects a specific movement to inspect.
-            ///
-            /// - Parameter movement: The selected movement type.
-            case openExerciseInfo(any MovementType)
-            
-            ///
-            case openMovementDetails(any MovementType, [any WorkoutSessionProtocol])
-            
-        }
-        
-        // MARK: - Destination
-        
-        /// Destination case for navigation
-        case destination(PresentationAction<Destination.Action>)
-    }
-    
-}
-
-
-import ComposableArchitecture
-import Foundation
-
-/// Represents the state for `ScoresFeature`
-extension ScoresFeature {
-    
-    @ObservableState
-    struct State {
-        
-        // MARK: - Properties
-        
-        ///
-        var selectedWorkout: NewGroupedWorkouts
-        
-        ///
-        var bestResult: (any WorkoutSessionProtocol)?
-        
-        // MARK: - Destination
-        
-        /// Represents the navigation destination state within `SummaryFeature`.
-        @Presents var destination: Destination.State?
-    }
-    
-}
-
-
-import ComposableArchitecture
-import Foundation
-
-/// Implementation of `ScoresFeature` destination
-extension ScoresFeature {
-    
-    @Reducer
-    enum Destination {
-        
-        /// Represents the destination for displaying in `MovementDetailsFeature`
-        case showDetails(MovementDetailsFeature)
-        
-        /// Represents the destination for displaying in `MovementInfoFeature`.
-        case openInfo(MovementInfoFeature)
-    }
-    
-}
 
 import ComposableArchitecture
 import SwiftUI
@@ -165,6 +14,7 @@ struct ScoresView: View {
     // MARK: - Properties
     
     @Bindable var store: StoreOf<ScoresFeature>
+    
     // MARK: - View
     
     var body: some View {
@@ -183,12 +33,15 @@ struct ScoresView: View {
             MovementInfoView(store: store)
                 .presentationDetents([.medium, .large])
         })
+        .sheet(item: $store.scope(state: \.destination?.openGoal, action: \.destination.openGoal), content: { store in
+            SetEditGoalView(store: store)
+        })
         .navigationDestination(
             item: $store.scope(
                 state: \.destination?.showDetails,
                 action: \.destination.showDetails)) { store in
-                MovementDetailsView(store: store)
-            }
+                    MovementDetailsView(store: store)
+                }
     }
     
     // MARK: - SubView
@@ -197,7 +50,7 @@ struct ScoresView: View {
     var toolbarButton: some ToolbarContent {
         ToolbarItem(placement: .topBarTrailing) {
             Button {
-                
+                send(.goalsButtonTapped)
             } label: {
                 Image(systemName: "plus")
             }
@@ -275,7 +128,7 @@ struct ScoresView: View {
     
     private func createCellScoreView(_ image: String, _ title: String, _ value: String?, _ date: Date?) -> some View {
         GeometryReader { geometry in
-            VStack { 
+            VStack {
                 HStack {
                     HStack {
                         Image(systemName: image)
